requirements.txt
python-telegram-bot==20.8
supabase==2.4.6
python-dotenv==1.0.0
